feat(wren-ui): rename promptThread to thread
