+++ wren-ui/src/components/pages/home/thread/CollapseContent.tsx
