+++ wren-ui/src/components/pages/home/thread/index.tsx
@@ -13,10 +13,7 @@ interface Props {
   onOpenSaveAsViewModal: (data: { sql: string; responseId: number }) => void;
 }
 
-const StyledPromptThread = styled.div`
-  width: 768px;
-  margin-left: auto;
-  margin-right: auto;
+const StyledThread = styled.div`
 
   h4.ant-typography {
     margin-top: 10px;
@@ -78,7 +75,7 @@ const AnswerResultTemplate = ({
 
 const AnswerResultIterator = makeIterable(AnswerResultTemplate);
 
-export default function PromptThread(props: Props) {
+export default function Thread(props: Props) {
   const { data, onOpenSaveAsViewModal } = props;
   const divRef = useRef<HTMLDivElement>(null);
 
@@ -105,12 +102,12 @@ export default function PromptThread(props: Props) {
   }, [divRef, data]);
 
   return (
-    <StyledPromptThread className="mt-12" ref={divRef}>
+    <StyledThread className="mt-12" ref={divRef}>
       <AnswerResultIterator
         data={data?.responses || []}
         onOpenSaveAsViewModal={onOpenSaveAsViewModal}
         onTriggerScrollToBottom={triggerScrollToBottom}
       />
-    </StyledPromptThread>
+    </StyledThread>
   );
 }
